+++ src/main/kotlin/ee/carlrobert/codegpt/codecompletions/CodeCompletionInsertAction.kt
@@ -31,6 +31,12 @@ class CodeCompletionInsertAction :
                 .map { it.element as CodeCompletionTextElement }
 
             if (elements.isEmpty()) {
+                val textToInsert = context.textToInsert()
+                val remainingCompletion = REMAINING_EDITOR_COMPLETION.get(editor) ?: ""
+                if (remainingCompletion.isNotEmpty()) {
+                    REMAINING_EDITOR_COMPLETION.set(editor, remainingCompletion.removePrefix(textToInsert))
+                }
+
                 InlineCompletion.getHandlerOrNull(editor)?.insert()
                 return
             }
@@ -80,7 +86,7 @@ class CodeCompletionInsertAction :
             val endOffset = element.textRange.endOffset
             editor.caretModel.moveToOffset(endOffset)
 
-            val remainingCompletionLine = REMAINING_EDITOR_COMPLETION.get(editor)
+            val remainingCompletionLine = (REMAINING_EDITOR_COMPLETION.get(editor) ?: "")
                 .removePrefix(element.text)
 
             processRemainingCompletion(remainingCompletionLine, editor, endOffset)
@@ -91,7 +97,7 @@ class CodeCompletionInsertAction :
             val lineEndOffset = editor.document.getLineEndOffset(lineNumber)
             editor.caretModel.moveToOffset(lineEndOffset)
 
-            val remainingText = REMAINING_EDITOR_COMPLETION.get(editor)
+            val remainingText = REMAINING_EDITOR_COMPLETION.get(editor) ?: ""
             val remainingCompletionLine = if (element.originalText.length > remainingText.length) {
                 remainingText.removePrefix(element.text)
             } else {