+++ src/main/kotlin/ee/carlrobert/codegpt/codecompletions/CodeCompletionSuggestionUpdateAdapter.kt
@@ -7,9 +7,19 @@ import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSug
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSuggestionUpdateManager.UpdateResult.Changed
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionSuggestionUpdateManager.UpdateResult.Invalidated
 import com.intellij.codeInsight.inline.completion.suggestion.InlineCompletionVariant
+import com.intellij.openapi.editor.Editor
+import ee.carlrobert.codegpt.CodeGPTKeys.REMAINING_EDITOR_COMPLETION
 
 class CodeCompletionSuggestionUpdateAdapter :
     InlineCompletionSuggestionUpdateManager.Default() {
+
+    override fun onDocumentChange(
+        event: InlineCompletionEvent.DocumentChange,
+        variant: InlineCompletionVariant.Snapshot
+    ): UpdateResult {
+        updateRemainingCompletion(event.editor, event.typing.typed)
+        return super.onDocumentChange(event, variant)
+    }
 
     override fun onCustomEvent(
         event: InlineCompletionEvent,
@@ -23,10 +33,20 @@ class CodeCompletionSuggestionUpdateAdapter :
         val textToInsert = event.toRequest().run {
             CompletionSplitter.split(completionText)
         }
+
+        updateRemainingCompletion(event.toRequest().editor, textToInsert)
+
         return Changed(
             variant.copy(
                 listOf(InlineCompletionGrayTextElement(completionText.removePrefix(textToInsert)))
             )
         )
+    }
+
+    private fun updateRemainingCompletion(editor: Editor, textToInsert: String) {
+        val remainingCompletion = REMAINING_EDITOR_COMPLETION.get(editor) ?: ""
+        if (remainingCompletion.isNotEmpty()) {
+            REMAINING_EDITOR_COMPLETION.set(editor, remainingCompletion.removePrefix(textToInsert))
+        }
     }
 }