+++ src/main/kotlin/ee/carlrobert/codegpt/settings/prompts/PromptsConfigurable.kt
@@ -1,6 +1,7 @@
 package ee.carlrobert.codegpt.settings.prompts
 
 import com.intellij.openapi.options.Configurable
+import ee.carlrobert.codegpt.actions.editor.EditorActionsUtil
 import ee.carlrobert.codegpt.settings.prompts.form.PromptsForm
 import javax.swing.JComponent
 
@@ -21,6 +22,7 @@ class PromptsConfigurable : Configurable {
 
     override fun apply() {
         component.applyChanges()
+        EditorActionsUtil.refreshActions()
     }
 
     override fun reset() {